Move all send notifications code
