+++ study/api-src/org/labkey/api/study/StudyInternalService.java
@@ -44,9 +44,6 @@ public interface StudyInternalService
 
     void generateNeededAlternateParticipantIds(Study study, User user);
 
-    @Migrate
-    void sendNewRequestNotifications(ViewContext context, SpecimenRequest request, BindException errors) throws Exception;
-
     void setLastSpecimenRequest(Study study, Integer lastSpecimenRequest);
 
     Integer getLastSpecimenRequest(Study study);
